--- vy6j3f(1)/src/App.js
@@ -1,3 +1,0 @@
-export default function Square() {
-  return <button className="square">X</button>;
-}
